@@ -8,13 +8,17 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "@popperjs/core": "^2.11.6",
+    "@vue/cli": "^5.0.8",
     "axios": "^0.23.0",
     "bootstrap": "^5.1.3",
     "bootstrap-vue": "^2.21.2",
+    "bootstrap-vue-3": "^0.3.3",
     "core-js": "^3.6.5",
-    "vue": "^2.6.14",
-    "vue-router": "^3.5.2",
-    "vuex": "^3.6.2"
+    "vue": "^3.2.31",
+    "vue-cli": "^2.9.6",
+    "vue-router": "^4.0.2",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
