@@ -68,6 +68,24 @@ module.exports = [
     },
   },
   {
+    name: "testCaseFramework",
+    type: "list",
+    message: "Select the Test Case Framework",
+    choices: [
+      { name: "MochaJS", value: "mochaJS" },
+      { name: "Jest", value: "jest" },
+      { name: "Jasmine", value: "jasmine" },
+      { name: "Karma", value: "karma" },
+      { name: "Puppeteer (Node Library)", value: "puppeteer" },
+      { name: "NightwatchJS", value: "nightwatchJS" },
+      { name: "Cypress", value: "cypress" },
+      { name: "None", value: false },
+    ],
+    when: (answers) => {
+      return answers.frontEnd;
+    },
+  },
+  {
     name: "theme",
     type: "list",
     message: "Do you want Theme-provider?",
@@ -76,7 +94,7 @@ module.exports = [
       { name: "None", value: false },
     ],
     when: (answers) => {
-      return answers.frontEndChoice === "react" || answers.frontEndChoice === "vue";
+      return answers.frontEnd;
     },
   },
   {
@@ -89,7 +107,7 @@ module.exports = [
       { name: "Okta", value: "Okta" },
       { name: "None", value: false },
     ],
-    when: (answers) => answers.frontEndChoice == "react",
+    when: (answers) => answers.frontEndChoice == "react" || answers.frontEndChoice === "vue",
   },
   {
     name: "store",
