@@ -3,17 +3,17 @@ import {React<% if(isAuth0){%> ,useEffect <%}%><% if(isDark && isMaterialUI){%> 
 <% if(isDark && isMaterialUI){%>import { ThemeProvider, createTheme } from '@mui/material/styles'<%}%>
 <% if(isAuth0){%>import { useAuth0 } from './react-spa';<%}%>
 
-<% if(!isOkta){ %> import NavBar from "./components/organisms/NavBar";<%}%>
+import NavBar from "./components/organisms/NavBar";
 <% if(isOkta){ %>import { BrowserRouter as Router } from 'react-router-dom';
-  import AppWithRouterAccess from './oktaFiles/AppWithRouterAccess';
+ 
 <%}%>
 import "./App.css";
 
-<%if(!isOkta && !isAuth0){%>
+
 const linksOfNav = [
   { href: '/home', label: 'Home' },
   { href: '/users', label: 'Users' },
-]<%}%>
+]
 
 const App = () => {
   <% if(isDark && isMaterialUI){%>
@@ -38,14 +38,16 @@ const App = () => {
       <%if(!isDark) {%>
         <%if(isOkta) {%>
           <Router>
-            <AppWithRouterAccess />
+            <NavBar brand='Made in India' links={linksOfNav}/>
           </Router>
         <%} else if(isAuth0) {%>
-          isUserAuthenticated && (<NavBar brand='Made in India'
-			              links={[
-			 	              { href: '/home', label: 'Home' },
-			 		            { href: '/users', label: 'Users' },
-			              ]}/>)
+          <>
+            {
+            isUserAuthenticated && (<NavBar brand='Made in India'
+            links={linksOfNav}/>)
+            }
+          </>
+
         <%} else {%>
           <NavBar brand='Made in India' links={linksOfNav}/>
         <%}%>
@@ -56,17 +58,24 @@ const App = () => {
       <%if(isDark && isMaterialUI) {%>
           <% if(isOkta) {%>
             <Router>
-              <AppWithRouterAccess />
+            <ThemeProvider theme={theme}>
+              <NavBar brand='Made in India'
+              links={linksOfNav} 
+              mode={mode}
+              setMode={setMode}/>
+              </ThemeProvider>
             </Router>
           <%} else if(isAuth0) { %>
+            <>
+              {
                   isUserAuthenticated && ((<ThemeProvider theme={theme}><NavBar brand='Made in India'
-				              links={[
-					              { href: '/home', label: 'Home' },
-				                { href: '/users', label: 'Users' },
-				              ]}
+                  links={linksOfNav}
                       mode={mode}
                       setMode={setMode}/>
                       </ThemeProvider>))
+              }
+            </>
+  
           <%} else { %>
                    <ThemeProvider theme={theme}>
                       <NavBar
@@ -83,17 +92,17 @@ const App = () => {
       <%if(isDark && !isMaterialUI) {%>
         <% if(isOkta) {%>
           <Router>
-            <AppWithRouterAccess />
+          <NavBar brand='Made in India' links={linksOfNav}/>
           </Router>
         <%} else if(isAuth0) { %>
+                <>
+                {
                 isUserAuthenticated && ((<NavBar brand='Made in India'
-                    links={[
-                      { href: '/home', label: 'Home' },
-                      { href: '/users', label: 'Users' },
-                    ]}/>
+                links={linksOfNav}/>
                     ))
+                  }
+                    </>
         <%} else { %>
-                 
                     <NavBar
                       brand='Made in India'
                         links={linksOfNav}/>
