@@ -1,5 +1,5 @@
 const fs = require("fs");
-const { updatePackage } = require("./helper");
+const { updateProjectDependencies } = require("./helper");
 const path = require("path");
 //Function to create logger service ------------------------------------------------------------>
 function createLogger(utilpath, loggerName, loggerTemplatePath, defaultRoute) {
@@ -7,7 +7,7 @@ function createLogger(utilpath, loggerName, loggerTemplatePath, defaultRoute) {
     let servicePath = path.join(utilpath, "utils", "logger");
     fs.mkdirSync(servicePath);
     let package = { name: "winston", version: "^3.3.3" };
-    updatePackage(utilpath, package);
+    updateProjectDependencies(utilpath, package);
     let contents = fs.readFileSync(
       loggerTemplatePath + "/" + loggerName + ".js",
       "utf-8"
@@ -17,7 +17,7 @@ function createLogger(utilpath, loggerName, loggerTemplatePath, defaultRoute) {
     });
   } else {
     let package = { name: "raven", version: "^2.6.4" };
-    updatePackage(utilpath, package);
+    updateProjectDependencies(utilpath, package);
   }
 }
   module.exports=createLogger
