@@ -159,14 +159,29 @@ function createDirectoryContents(
 }
 
 //to update package.json------------------------------------------------>
-function updatePackage(path, package) {
+function updateProjectDependencies(path, dependency) {
   let packagefile = fs.readFileSync(`${path}/package.json`, "utf-8");
   packagefile = JSON.parse(packagefile);
   let newPackageFile = {
     ...packagefile,
     dependencies: {
       ...packagefile.dependencies,
-      [package.name]: package.version,
+      [dependency.name]: dependency.version,
+    },
+  };
+  newPackageFile = JSON.stringify(newPackageFile);
+  fs.writeFileSync(`${path}/package.json`, newPackageFile, "utf-8");
+}
+
+// To update the scripts in Package.json
+function updateProjectScripts(path, script) {
+  let packagefile = fs.readFileSync(`${path}/package.json`, "utf-8");
+  packagefile = JSON.parse(packagefile);
+  let newPackageFile = {
+    ...packagefile,
+    scripts: {
+      ...packagefile.scripts,
+      [script.name]: script.command,
     },
   };
   newPackageFile = JSON.stringify(newPackageFile);
@@ -175,5 +190,6 @@ function updatePackage(path, package) {
 
 module.exports = {
   createDirectoryContents,
-  updatePackage,
+  updateProjectDependencies,
+  updateProjectScripts
 };
